# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  id: string;
  title: string;
  author: string;
  subject: string;
  classLevel: number;
  price: number;
  originalPrice: number;
  imageUrl: string;
  description: string;
  rating: number;
  inStock: boolean;
  /** @nullable */
  badge?: string | null;
}

export interface Promotion {
  id: string;
  eyebrow: string;
  title: string;
  subtitle: string;
  discount: number;
  imageUrl: string;
  background: string;
}

export interface ClassFilter {
  level: number;
  label: string;
  bookCount: number;
}

export interface SubjectFilter {
  name: string;
  bookCount: number;
}

export interface CatalogFilters {
  classes: ClassFilter[];
  subjects: SubjectFilter[];
}

export interface CatalogSummary {
  bookCount: number;
  classCount: number;
  subjectCount: number;
  promotionCount: number;
}

export interface CartLine {
  book: Book;
  quantity: number;
  lineTotal: number;
}

export interface Cart {
  items: CartLine[];
  itemCount: number;
  subtotal: number;
}

export interface CartItemInput {
  bookId: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  quantity: number;
}

export interface CartItemUpdate {
  /**
     * @minimum 1
     * @maximum 20
     */
  quantity: number;
}

export interface OrderLine {
  bookId: string;
  title: string;
  imageUrl: string;
  unitPrice: number;
  quantity: number;
  lineTotal: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  placed: 'placed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderPaymentMethod = typeof OrderPaymentMethod[keyof typeof OrderPaymentMethod];


export const OrderPaymentMethod = {
  cash_on_delivery: 'cash_on_delivery',
} as const;

export interface Order {
  id: string;
  status: OrderStatus;
  paymentMethod: OrderPaymentMethod;
  items: OrderLine[];
  itemCount: number;
  total: number;
  customerName: string;
  phone: string;
  addressLine: string;
  city: string;
  state: string;
  postalCode: string;
  createdAt: string;
}

export type CheckoutInputPaymentMethod = typeof CheckoutInputPaymentMethod[keyof typeof CheckoutInputPaymentMethod];


export const CheckoutInputPaymentMethod = {
  cash_on_delivery: 'cash_on_delivery',
} as const;

export interface CheckoutInput {
  /** @minLength 2 */
  customerName: string;
  /**
     * @minLength 10
     * @maxLength 16
     */
  phone: string;
  /** @minLength 5 */
  addressLine: string;
  /** @minLength 2 */
  city: string;
  /** @minLength 2 */
  state: string;
  /**
     * @minLength 5
     * @maxLength 10
     */
  postalCode: string;
  paymentMethod: CheckoutInputPaymentMethod;
}

export type ListBooksParams = {
q?: string;
/**
 * @minimum 1
 * @maximum 12
 */
classLevel?: number;
subject?: string;
dealOnly?: boolean;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};