# packages/api-zod/src/generated/types/book.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Book {
  id: string;
  title: string;
  author: string;
  subject: string;
  classLevel: number;
  price: number;
  originalPrice: number;
  imageUrl: string;
  description: string;
  rating: number;
  inStock: boolean;
  /** @nullable */
  badge?: string | null;
}