# packages/api-zod/src/generated/types/cart.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CartLine } from './cartLine';

export interface Cart {
  items: CartLine[];
  itemCount: number;
  subtotal: number;
}