# packages/api-zod/src/generated/types/cartItemUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemUpdate {
  /**
     * @minimum 1
     * @maximum 20
     */
  quantity: number;
}