# packages/api-zod/src/generated/types/cartLine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from './book';

export interface CartLine {
  book: Book;
  quantity: number;
  lineTotal: number;
}