# packages/api-zod/src/generated/types/catalogFilters.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClassFilter } from './classFilter';
import type { SubjectFilter } from './subjectFilter';

export interface CatalogFilters {
  classes: ClassFilter[];
  subjects: SubjectFilter[];
}