# packages/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutInputPaymentMethod } from './checkoutInputPaymentMethod';

export interface CheckoutInput {
  /** @minLength 2 */
  customerName: string;
  /**
     * @minLength 10
     * @maxLength 16
     */
  phone: string;
  /** @minLength 5 */
  addressLine: string;
  /** @minLength 2 */
  city: string;
  /** @minLength 2 */
  state: string;
  /**
     * @minLength 5
     * @maxLength 10
     */
  postalCode: string;
  paymentMethod: CheckoutInputPaymentMethod;
}