# packages/api-zod/src/generated/types/classFilter.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClassFilter {
  level: number;
  label: string;
  bookCount: number;
}