# packages/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLine } from './orderLine';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  status: OrderStatus;
  paymentMethod: OrderPaymentMethod;
  items: OrderLine[];
  itemCount: number;
  total: number;
  customerName: string;
  phone: string;
  addressLine: string;
  city: string;
  state: string;
  postalCode: string;
  createdAt: Date;
}