# packages/api-zod/src/generated/types/promotion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Promotion {
  id: string;
  eyebrow: string;
  title: string;
  subtitle: string;
  discount: number;
  imageUrl: string;
  background: string;
}